- Split settings responsibilities: moved preset functions to a dedicated `presets-service.js`. - Updated imports to point to the new service. - Because who doesn’t love a good refactor that screams “separation of concerns”?
